# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "panda-wiki API documentation",
        "title": "panda-wiki API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/model": {
            "put": {
                "description": "update model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "model"
                ],
                "parameters": [
                    {
                        "description": "update model request",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateModelReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "create model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "model"
                ],
                "summary": "create model",
                "parameters": [
                    {
                        "description": "create model request",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateModelReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/model/auto-mode": {
            "post": {
                "description": "update BaiZhiCloud API key and optional chat model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "model"
                ],
                "summary": "update BaiZhiCloud model setting",
                "parameters": [
                    {
                        "description": "update BaiZhiCloud model setting request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateAutoModelSettingReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/model/check": {
            "post": {
                "description": "check model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "model"
                ],
                "summary": "check model",
                "parameters": [
                    {
                        "description": "check model request",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_chaitin_panda-wiki_domain.CheckModelReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_chaitin_panda-wiki_domain.CheckModelResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/v1/model/provider/supported": {
            "post": {
                "description": "get provider supported model list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "model"
                ],
                "summary": "get provider supported model list",
                "parameters": [
                    {
                        "description": "get supported model list request",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.GetProviderModelListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.PWResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.GetProviderModelListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/model/switch-mode": {
            "post": {
                "description": "switch model mode between manual and auto",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "model"
                ],
                "summary": "switch mode",
                "parameters": [
                    {
                        "description": "switch mode request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.SwitchModeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.SwitchModeResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.CreateModelReq": {
            "type": "object",
            "required": [
                "base_url",
                "model",
                "provider",
                "type"
            ],
            "properties": {
                "api_header": {
                    "type": "string"
                },
                "api_key": {
                    "type": "string"
                },
                "api_version": {
                    "description": "for azure openai",
                    "type": "string"
                },
                "base_url": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "parameters": {
                    "$ref": "#/definitions/github_com_chaitin_panda-wiki_domain.ModelParam"
                },
                "provider": {
                    "$ref": "#/definitions/github_com_chaitin_panda-wiki_domain.ModelProvider"
                },
                "type": {
                    "enum": [
                        "chat",
                        "embedding",
                        "rerank",
                        "analysis",
                        "analysis-vl"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.ModelType"
                        }
                    ]
                }
            }
        },
...
        "domain.GetProviderModelListReq": {
            "type": "object",
            "required": [
                "base_url",
                "provider",
                "type"
            ],
            "properties": {
                "api_header": {
                    "type": "string"
                },
                "api_key": {
                    "type": "string"
                },
                "base_url": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "type": {
                    "enum": [
                        "chat",
                        "embedding",
                        "rerank",
                        "analysis",
                        "analysis-vl"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.ModelType"
                        }
                    ]
                }
            }
        },
        "domain.GetProviderModelListResp": {
            "type": "object",
            "properties": {
                "models": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ProviderModelListItem"
                    }
                }
            }
        },
...
        "domain.ModelType": {
            "type": "string",
            "enum": [
                "chat",
                "embedding",
                "rerank",
                "analysis",
                "analysis-vl"
            ],
            "x-enum-varnames": [
                "ModelTypeChat",
                "ModelTypeEmbedding",
                "ModelTypeRerank",
                "ModelTypeAnalysis",
                "ModelTypeAnalysisVL"
            ]
        },
...
        "domain.PWResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "domain.ProviderModelListItem": {
            "type": "object",
            "properties": {
                "model": {
                    "type": "string"
                }
            }
        },
...
        "domain.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "domain.SwitchModeReq": {
            "type": "object",
            "required": [
                "mode"
            ],
            "properties": {
                "mode": {
                    "type": "string",
                    "enum": [
                        "manual",
                        "auto"
                    ]
                }
            }
        },
        "domain.SwitchModeResp": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "domain.UpdateAutoModelSettingReq": {
            "type": "object",
            "required": [
                "APIKey"
            ],
            "properties": {
                "APIKey": {
                    "description": "APIKey 百智云 API 密钥",
                    "type": "string"
                },
                "ChatModel": {
                    "description": "ChatModel 对话模型，允许为空表示清空为默认",
                    "type": "string"
                }
            }
        },
...
        "domain.UpdateModelReq": {
            "type": "object",
            "required": [
                "base_url",
                "id",
                "model",
                "provider",
                "type"
            ],
            "properties": {
                "api_header": {
                    "type": "string"
                },
                "api_key": {
                    "type": "string"
                },
                "api_version": {
                    "description": "for azure openai",
                    "type": "string"
                },
                "base_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "model": {
                    "type": "string"
                },
                "parameters": {
                    "$ref": "#/definitions/github_com_chaitin_panda-wiki_domain.ModelParam"
                },
                "provider": {
                    "$ref": "#/definitions/github_com_chaitin_panda-wiki_domain.ModelProvider"
                },
                "type": {
                    "enum": [
                        "chat",
                        "embedding",
                        "rerank",
                        "analysis",
                        "analysis-vl"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.ModelType"
                        }
                    ]
                }
            }
        },
...
        "github_com_chaitin_panda-wiki_domain.CheckModelReq": {
            "type": "object",
            "required": [
                "base_url",
                "model",
                "provider",
                "type"
            ],
            "properties": {
                "api_header": {
                    "type": "string"
                },
                "api_key": {
                    "type": "string"
                },
                "api_version": {
                    "description": "for azure openai",
                    "type": "string"
                },
                "base_url": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "parameters": {
                    "$ref": "#/definitions/github_com_chaitin_panda-wiki_domain.ModelParam"
                },
                "provider": {
                    "$ref": "#/definitions/github_com_chaitin_panda-wiki_domain.ModelProvider"
                },
                "type": {
                    "enum": [
                        "chat",
                        "embedding",
                        "rerank",
                        "analysis",
                        "analysis-vl"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.ModelType"
                        }
                    ]
                }
            }
        },
        "github_com_chaitin_panda-wiki_domain.CheckModelResp": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "github_com_chaitin_panda-wiki_domain.ModelParam": {
            "type": "object",
            "properties": {
                "context_window": {
                    "type": "integer"
                },
                "max_tokens": {
                    "type": "integer"
                },
                "r1_enabled": {
                    "type": "boolean"
                },
                "support_computer_use": {
                    "type": "boolean"
                },
                "support_images": {
                    "type": "boolean"
                },
                "support_prompt_cache": {
                    "type": "boolean"
                },
                "temperature": {
                    "type": "number"
                }
            }
        },
        "github_com_chaitin_panda-wiki_domain.ModelProvider": {
            "type": "string",
            "enum": [
                "BaiZhiCloud"
            ],
            "x-enum-varnames": [
                "ModelProviderBrandBaiZhiCloud"
            ]
        },
...
    },
    "securityDefinitions": {
        "bearerAuth": {
            "description": "Type \"Bearer\" + a space + your token to authorize",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}